Move GeoJSONReader
